backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
requests
python-dotenv
pydantic
